+++ app/src/main/java/com/example/huangshan/common/ui/LoginFragment.java
@@ -171,7 +171,7 @@ public class LoginFragment extends Fragment implements View.OnClickListener{
         SharedPreferences preferences= getActivity().getSharedPreferences("loginUser", MODE_PRIVATE);
         SharedPreferences.Editor editor = preferences.edit();
         editor.putString("role", "admin");
-        editor.putInt("id",admin.getId());
+        editor.putLong("id",admin.getId());
         editor.putString("account",admin.getAccount());
         editor.putString("name",admin.getName());
         editor.putString("sex",admin.getSex());
@@ -193,7 +193,7 @@ public class LoginFragment extends Fragment implements View.OnClickListener{
         SharedPreferences preferences= getActivity().getSharedPreferences("loginUser", MODE_PRIVATE);
         SharedPreferences.Editor editor = preferences.edit();
         editor.putString("role", "tourist");
-        editor.putInt("id",tourist.getId());
+        editor.putLong("id",tourist.getId());
         editor.putString("account",tourist.getAccount());
         editor.putString("name",tourist.getName());
         editor.putString("sex",tourist.getSex());
